backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
idna==3.10
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
redis==6.2.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13

# Production server
gunicorn==23.0.0

# PostgreSQL adapter
psycopg2-binary==2.9.10

# Excel generation
openpyxl==3.1.5
Pillow==10.4.0  # Required for image support in Excel files

# Development dependencies (optional, for hot-reload in containers)
watchdog==6.0.0
